refactor(server): extract marketing/SEO SSR → src/server/marketing.js
Pure move, zero behavior change. Pulls the public marketing-page
server-side rendering cluster out of server.js into src/server/marketing.js
(~159 lines).

Why: continues the server.js dismemberment. This is a fully self-contained
unit — static SEO metadata + one pure render function, no DB/network/LLM.

What:
- New src/server/marketing.js holds FAQ_PAIRS, FAQ_BODY_HTML, FAQ_JSON_LD,
  MARKETING_META, ORG_JSON_LD, WEBSITE_JSON_LD, DEFAULT_OG_IMAGE, and
  renderMarketingPage. Only MARKETING_META + renderMarketingPage are
  exported (the two the /, /product, /faq SSR route consumes); the FAQ
  content, JSON-LD blocks, and OG image stay module-private (all were
  internal-only — every one was def + a single in-cluster use).
- server.js imports the two and keeps the SSR route in place; the inline
  block is replaced with a breadcrumb comment. Code moved verbatim.

Test plan:
- node --check server.js + marketing.js -> OK
- npm run lint (no-undef gate) -> clean (confirms the two imported symbols
  resolve and no orphaned reference to the now-private consts remains)
- route-inventory guard -> PASS (snapshot unchanged)
- vitest -> 78/78 pass (+5 new marketing tests: META covers the 3 routes;
  renderMarketingPage injects title/canonical/Org+WebSite JSON-LD, FAQ
  JSON-LD only on /faq, crawler body inside #root, ogImage override vs
  default card)

Rollback: revert; pure copy, restores the inline block exactly.
